+++ src/java/boa/datagen/SeqRepoImporter.java
@@ -123,7 +123,7 @@ public class SeqRepoImporter {
 							continue;
 						// new model
 						if (!STORE_ASTS) {
-							int size = MAX_SIZE_FOR_PROJECT_WITH_COMMITS * DefaultProperties.MAX_SIZE_FACTOR;
+							double size = MAX_SIZE_FOR_PROJECT_WITH_COMMITS * DefaultProperties.MAX_SIZE_FACTOR;
 							int curSize = repo.getSize() * 1000;
 							if (curSize > size) {
 								DefaultProperties.exceptions.put(repo.name, "repo size: " + curSize + " exceeding the max size: " + size);
@@ -385,7 +385,7 @@ public class SeqRepoImporter {
 			if (!STORE_ASTS) {
 				ByteArrayFile f = new ByteArrayFile(gitDir.getAbsolutePath());
 				BytesWritable bw = new BytesWritable(SerializationUtils.serialize(f));
-				int size =  MAX_SIZE_FOR_PROJECT_WITH_COMMITS * DefaultProperties.MAX_SIZE_FACTOR;
+				double size =  MAX_SIZE_FOR_PROJECT_WITH_COMMITS * DefaultProperties.MAX_SIZE_FACTOR;
 				if (!f.isBuilt() || bw.getLength() > size) {
 					DefaultProperties.exceptions.put(name, "pack file size: " + bw.getLength() + " exceeding the max size: " + size);
 					noast = true;
